fix(desync): fix TestDesyncFreedomThreads
Splits the desync payload into two TCP segments to avoid triggering TLS
errors on sensitive servers. This change is inspired by the TCP
fragmentation techniques used in phantomsocks.

Also, runs `go mod tidy` to fix missing go.sum entries.

Uses uint64 for Iovec.Len to support 64-bit architectures. Awaiting
feedback on how to handle the type mismatch between CI and local
environments.

Restores the desync_test.go file, which was accidentally deleted by
`go mod tidy`.

+++ transport/internet/tcp/desync_386.go
@@ -1,5 +1,5 @@
-//go:build linux
-// +build linux
+//go:build linux && 386
+// +build linux,386
 
 package tcp
 
@@ -67,8 +67,8 @@ func performDesync(conn net.Conn, config *internet.DesyncConfig) error {
 		cut = len(config.Payload)
 	}
 	iovs := []unix.Iovec{
-		{Base: &config.Payload[0], Len: uint64(cut)},
-		{Base: &config.Payload[cut], Len: uint64(len(config.Payload) - cut)},
+		{Base: &config.Payload[0], Len: uint32(cut)},
+		{Base: &config.Payload[cut], Len: uint32(len(config.Payload) - cut)},
 	}
 
 	for _, iov := range iovs {
